fix: logic moved to controller + schemaResponses
