+++ src/event/dto/get-statistics-by-id-event.dto.ts
@@ -1,16 +1,27 @@
+import { ticketSchema } from '@/ticket/dto/ticket.dto';
 import z from 'zod';
+import { eventTicketsSchema } from './event-tickets.dto';
+import { eventSchema } from './event.dto';
 
-export const getStatisticsByIdResponseSchema = z.object({
-  test: z.string(),
-  // event: eventSchema,
-  // eventTickets: z.object({
-  //   eventTickets: z.array(
-  //       baseEventTicketsSchema.pick({
-  //         id: true,
-  //         amount: true,
-  //         type: true,
-  //         price: true,
-  //       }),
-  //   ),
-  // })
-});
+export const getStatisticsByIdResponseSchema = eventSchema.merge(
+  z.object({
+    tickets: z.array(ticketSchema),
+    eventTickets: z.array(eventTicketsSchema),
+    statistics: z
+      .object({
+        maxTickets: z.number(),
+        emmitedTickets: z.number(),
+        emittedTicketsPercent: z.number(),
+        emmitedticketPerType: z.object({}),
+        totalIncome: z.number(),
+        maxTotalIncome: z.number(),
+        maxTicketPerType: z.object({}),
+        totalTicketsScanned: z.number(),
+        notScanned: z.number(),
+        attendancePercent: z.number(),
+        attendancePerHour: z.object({}),
+        avgAmountPerTicketGroup: z.object({}).nullable(),
+      })
+      .optional(),
+  }),
+);
